Rename GitHub providers to match their IDs
