+++ update-package-via-manifest/src/providers/github-releases.ts
